Improved random number generation for multiplexing
Previously, a multiplexing operation would use a RNG initialized with
a seed from `std::random_device`, and this deterministic RNG would be
used for the entire operation.

With this modification, every X bytes the RNG is reset with another
random number from std::random_device (where X is an arbitrary constant,
currently set to 512), in order to mitigate (improbable?) cryptanalysis
techniques.